requirements.txt
alembic==1.10.3
Babel==2.12.1
bidict==0.22.1
blinker==1.6
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Babel==2.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2023.3
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post4
SQLAlchemy==2.0.5.post1
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
